#2 More refactoring
