+++ exasol/ai/mcp/server/mcp_server.py
@@ -1,5 +1,4 @@
 import json
-import logging
 import os
 import traceback
 from textwrap import dedent
@@ -7,114 +6,17 @@ from typing import Annotated
 
 from fastmcp import FastMCP
 from mcp.types import TextContent
-from pydantic import (
-    BaseModel,
-    Field,
-)
+from pydantic import Field
 from pyexasol import (
     ExaConnection,
     connect,
 )
 
-logger = logging.getLogger("exasol_mcp_server")
-
-
-class MetaSettings(BaseModel):
-    """
-    The settings for a single type of metadata, e.g. tables.
-    """
-
-    enable: bool = True
-    """
-    Allows to disable the listing of a particular type of metadata.
-    """
-
-    name_field: str = "name"
-    """
-    The name of the output field that contains the object name, e.g. "table_name".
-    """
-
-    comment_field: str = "comment"
-    """
-    The name of the output field that contains the comment, e.g. "table_comment".
-    """
-
-    like_pattern: str | None = None
-    """
-    An optional sql-style pattern for the object name filtering.
-
-    Use case example: The user wants to create a set of purified de-normalised views on
-    the existing database and limit the table listings to only these views. One way of
-    achieving this is to create the views in a new schema and limit the listing of the
-    schemas to this schema only. In the case of no permission to create schema, one can
-    create the views in an existing schema and use some prefix for name disambiguation.
-    This prefix can also be used for filtering the views in the listing.
-    """
-    regexp_pattern: str | None = None
-    """
-    An optional regular expression pattern for the object name filtering.
-    Both like_pattern and regexp_pattern can be used at the same time, although there is
-    not much point in doing so.
-    """
-
-    @property
-    def select_predicate(self) -> str:
-        """
-        The SQL predicate for the object filtering by name.
-        Empty string if neigher of the filtering patterns are defined.
-        """
-        conditions: list[str] = []
-        if self.like_pattern:
-            conditions.append(
-                f"""local."{self.name_field}" LIKE '{self.like_pattern}'"""
-            )
-        if self.regexp_pattern:
-            conditions.append(
-                f"""local."{self.name_field}" REGEXP_LIKE '{self.regexp_pattern}'"""
-            )
-        return " AND ".join(conditions)
-
-
-class MetaColumnSettings(MetaSettings):
-    """
-    The settings for listing columns when describing a table. Adds few more fields to
-    the metadata output.
-    """
-
-    type_field: str = "column_type"
-    primary_key_field: str = "primary_key"
-    foreign_key_field: str = "foreign_key"
-
-
-class McpServerSettings(BaseModel):
-    """
-    MCP server configuration.
-    """
-
-    schemas: MetaSettings = MetaSettings(
-        name_field="schema_name", comment_field="schema_comment"
-    )
-    tables: MetaSettings = MetaSettings(
-        name_field="table_name", comment_field="table_comment"
-    )
-    views: MetaSettings = MetaSettings(
-        enable=False, name_field="table_name", comment_field="table_comment"
-    )
-    functions: MetaSettings = MetaSettings(
-        name_field="function_name", comment_field="function_comment"
-    )
-    scripts: MetaSettings = MetaSettings(
-        name_field="script_name", comment_field="script_comment"
-    )
-    columns: MetaColumnSettings = MetaColumnSettings(
-        name_field="column_name", comment_field="column_comment"
-    )
-
-
-def _report_error(tool_name: str, error_message: str) -> TextContent:
-    logger.error("Error in %s: %s", tool_name, error_message)
-    error_json = json.dumps({"error": error_message})
-    return TextContent(type="text", text=error_json)
+from exasol.ai.mcp.server.server_settings import (
+    McpServerSettings,
+    MetaSettings,
+)
+from exasol.ai.mcp.server.utils import report_error
 
 
 def _where_clause(*predicates) -> str:
@@ -248,7 +150,7 @@ class ExasolMCPServer(FastMCP):
             result_json = json.dumps(result)
             return TextContent(type="text", text=result_json)
         except Exception:  # pylint: disable=broad-exception-caught
-            return _report_error(tool_name, traceback.format_exc())
+            return report_error(tool_name, traceback.format_exc())
 
     def list_schemas(self) -> TextContent:
         tool_name = self.list_functions.__name__
@@ -282,7 +184,7 @@ class ExasolMCPServer(FastMCP):
         tool_name = self.list_functions.__name__
         conf = self.config.functions
         if not conf.enable:
-            return _report_error(tool_name, "Function listing is disabled.")
+            return report_error(tool_name, "Function listing is disabled.")
 
         query = self._build_meta_query("FUNCTION", conf, schema_name)
         return self._execute_query(tool_name, query)
@@ -296,7 +198,7 @@ class ExasolMCPServer(FastMCP):
         tool_name = self.list_scripts.__name__
         conf = self.config.scripts
         if not conf.enable:
-            return _report_error(tool_name, "Script listing is disabled.")
+            return report_error(tool_name, "Script listing is disabled.")
 
         query = self._build_meta_query(
             "SCRIPT", conf, schema_name, "SCRIPT_TYPE = 'UDF'"
@@ -313,12 +215,12 @@ class ExasolMCPServer(FastMCP):
         tool_name = self.describe_table.__name__
         conf = self.config.columns
         if not conf.enable:
-            return _report_error(tool_name, "Column listing is disabled.")
+            return report_error(tool_name, "Column listing is disabled.")
         schema_name = schema_name or self.connection.current_schema()
         if not schema_name:
-            return _report_error(tool_name, "Schema name is not provided.")
+            return report_error(tool_name, "Schema name is not provided.")
         if not table_name:
-            return _report_error(tool_name, "Table name is not provided.")
+            return report_error(tool_name, "Table name is not provided.")
 
         c_predicates = [
             f"COLUMN_SCHEMA = '{schema_name}'",
