+++ exasol/ai/mcp/server/utils.py
@@ -1,0 +1,12 @@
+import json
+import logging
+
+from mcp.types import TextContent
+
+logger = logging.getLogger("exasol_mcp_server")
+
+
+def report_error(tool_name: str, error_message: str) -> TextContent:
+    logger.error("Error in %s: %s", tool_name, error_message)
+    error_json = json.dumps({"error": error_message})
+    return TextContent(type="text", text=error_json)
